+++ packages/auth-plugins/src/LoginForm.tsx
@@ -49,7 +49,7 @@ export function LoginForm({
           >
             {isLoggingIn && (
               <span>
-                <LoadingSpinner className="loading-spinner-vertical-align" />
+                <LoadingSpinner className="mr-2 loading-spinner-vertical-align" />
                 <span className="btn-normal-content">Logging in</span>
                 <span className="btn-hover-content">Cancel</span>
               </span>